+++ src/components/AccountDrawer/MiniPortfolio/ExpandoRow.tsx
@@ -30,6 +30,10 @@ const Wrapper = styled(Column)<{ numItems: number; isExpanded: boolean }>`
   overflow: hidden;
 `
 
+/*
+ * TODO(WEB-XXXX):
+ * Replace this component to use `components/Expand` under the hood
+ */
 type ExpandoRowProps = PropsWithChildren<{ title?: string; numItems: number; isExpanded: boolean; toggle: () => void }>
 export function ExpandoRow({ title = t`Hidden`, numItems, isExpanded, toggle, children }: ExpandoRowProps) {
   if (numItems === 0) return null
