+++ src/components/Settings/index.tsx
@@ -3,41 +3,26 @@ import { t, Trans } from '@lingui/macro'
 import { Percent } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
 import { sendEvent } from 'components/analytics'
+import Column, { AutoColumn } from 'components/Column'
+import { RowBetween, RowFixed } from 'components/Row'
+import Toggle from 'components/Toggle'
+import { L2_CHAIN_IDS } from 'constants/chains'
+import { useOnClickOutside } from 'hooks/useOnClickOutside'
 import { isSupportedChainId } from 'lib/hooks/routing/clientSideSmartOrderRouter'
-import { useRef, useState } from 'react'
-import { Settings, X } from 'react-feather'
-import { Text } from 'rebass'
-import styled, { useTheme } from 'styled-components/macro'
+import { useRef } from 'react'
+import { Settings } from 'react-feather'
+import { useModalIsOpen, useToggleSettingsMenu } from 'state/application/hooks'
+import { ApplicationModal } from 'state/application/reducer'
+import { useClientSideRouter } from 'state/user/hooks'
+import styled from 'styled-components/macro'
+import { ThemedText } from 'theme'
 
-import { useOnClickOutside } from '../../hooks/useOnClickOutside'
-import { useModalIsOpen, useToggleSettingsMenu } from '../../state/application/hooks'
-import { ApplicationModal } from '../../state/application/reducer'
-import { useClientSideRouter, useExpertModeManager } from '../../state/user/hooks'
-import { ThemedText } from '../../theme'
-import { ButtonError } from '../Button'
-import { AutoColumn } from '../Column'
-import Modal from '../Modal'
-import QuestionHelper from '../QuestionHelper'
-import { RowBetween, RowFixed } from '../Row'
-import Toggle from '../Toggle'
-import TransactionSettings from '../TransactionSettings'
+import MaxSlippageSettings from './MaxSlippageSettings'
+import TransactionDeadlineSettings from './TransactionDeadlineSettings'
 
 const StyledMenuIcon = styled(Settings)`
   height: 20px;
   width: 20px;
-
-  > * {
-    stroke: ${({ theme }) => theme.textSecondary};
-  }
-`
-
-const StyledCloseIcon = styled(X)`
-  height: 20px;
-  width: 20px;
-  :hover {
-    cursor: pointer;
-  }
-
   > * {
     stroke: ${({ theme }) => theme.textSecondary};
   }
@@ -53,7 +38,6 @@ const StyledMenuButton = styled.button<{ disabled: boolean }>`
   padding: 0;
   border-radius: 0.5rem;
   height: 20px;
-
   ${({ disabled }) =>
     !disabled &&
     `
@@ -64,12 +48,6 @@ const StyledMenuButton = styled.button<{ disabled: boolean }>`
       opacity: 0.7;
     }
   `}
-`
-const EmojiWrapper = styled.div`
-  position: absolute;
-  bottom: -6px;
-  right: 0px;
-  font-size: 14px;
 `
 
 const StyledMenu = styled.div`
@@ -97,89 +75,39 @@ const MenuFlyout = styled.span`
   right: 0rem;
   z-index: 100;
   color: ${({ theme }) => theme.textPrimary};
-
   ${({ theme }) => theme.deprecated_mediaWidth.deprecated_upToMedium`
     min-width: 18.125rem;
   `};
-
   user-select: none;
 `
 
-const Break = styled.div`
-  width: 100%;
-  height: 1px;
-  background-color: ${({ theme }) => theme.deprecated_bg3};
+const ClientSideRouterSettings = styled.div<{ enabled: boolean }>`
+  opacity: ${({ enabled }) => (enabled ? 1 : 0.3)};
+  pointer-events: ${({ enabled }) => (enabled ? 'auto' : 'none')};
 `
 
-const ModalContentWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 2rem 0;
-  background-color: ${({ theme }) => theme.backgroundInteractive};
-  border-radius: 20px;
+const Divider = styled.div`
+  width: 100%;
+  height: 1px;
+  border-width: 0;
+  margin: 0;
+  background-color: ${({ theme }) => theme.backgroundOutline};
 `
 
 export default function SettingsTab({ placeholderSlippage }: { placeholderSlippage: Percent }) {
   const { chainId } = useWeb3React()
+  const showDeadlineSettings = Boolean(chainId && !L2_CHAIN_IDS.includes(chainId))
 
   const node = useRef<HTMLDivElement | null>(null)
   const open = useModalIsOpen(ApplicationModal.SETTINGS)
+
   const toggle = useToggleSettingsMenu()
-
-  const theme = useTheme()
-
-  const [expertMode, toggleExpertMode] = useExpertModeManager()
-
-  const [clientSideRouter, setClientSideRouter] = useClientSideRouter()
-
-  // show confirmation view before turning on
-  const [showConfirmation, setShowConfirmation] = useState(false)
-
   useOnClickOutside(node, open ? toggle : undefined)
+
+  const [clientSideRouterEnabled, setClientSideRouterEnabled] = useClientSideRouter()
 
   return (
     <StyledMenu ref={node}>
-      <Modal isOpen={showConfirmation} onDismiss={() => setShowConfirmation(false)} maxHeight={100}>
-        <ModalContentWrapper>
-          <AutoColumn gap="lg">
-            <RowBetween style={{ padding: '0 2rem' }}>
-              <div />
-              <Text fontWeight={500} fontSize={20}>
-                <Trans>Are you sure?</Trans>
-              </Text>
-              <StyledCloseIcon onClick={() => setShowConfirmation(false)} />
-            </RowBetween>
-            <Break />
-            <AutoColumn gap="lg" style={{ padding: '0 2rem' }}>
-              <Text fontWeight={500} fontSize={20}>
-                <Trans>
-                  Expert mode turns off the confirm transaction prompt and allows high slippage trades that often result
-                  in bad rates and lost funds.
-                </Trans>
-              </Text>
-              <Text fontWeight={600} fontSize={20}>
-                <Trans>ONLY USE THIS MODE IF YOU KNOW WHAT YOU ARE DOING.</Trans>
-              </Text>
-              <ButtonError
-                error={true}
-                padding="12px"
-                onClick={() => {
-                  const confirmWord = t`confirm`
-                  if (window.prompt(t`Please type the word "${confirmWord}" to enable expert mode.`) === confirmWord) {
-                    toggleExpertMode()
-                    setShowConfirmation(false)
-                  }
-                }}
-              >
-                <Text fontSize={20} fontWeight={500} id="confirm-expert-mode">
-                  <Trans>Turn On Expert Mode</Trans>
-                </Text>
-              </ButtonError>
-            </AutoColumn>
-          </AutoColumn>
-        </ModalContentWrapper>
-      </Modal>
       <StyledMenuButton
         disabled={!isSupportedChainId(chainId)}
         onClick={toggle}
@@ -187,72 +115,47 @@ export default function SettingsTab({ placeholderSlippage }: { placeholderSlippa
         aria-label={t`Transaction Settings`}
       >
         <StyledMenuIcon data-testid="swap-settings-button" />
-        {expertMode ? (
-          <EmojiWrapper>
-            <span role="img" aria-label="wizard-icon">
-              🧙
-            </span>
-          </EmojiWrapper>
-        ) : null}
       </StyledMenuButton>
       {open && (
         <MenuFlyout>
-          <AutoColumn gap="md" style={{ padding: '1rem' }}>
-            <Text fontWeight={600} fontSize={14}>
-              <Trans>Settings</Trans>
-            </Text>
-            <TransactionSettings placeholderSlippage={placeholderSlippage} />
-            <Text fontWeight={600} fontSize={14}>
-              <Trans>Interface Settings</Trans>
-            </Text>
+          <AutoColumn gap="16px" style={{ padding: '1rem' }}>
             {isSupportedChainId(chainId) && (
               <RowBetween>
                 <RowFixed>
-                  <ThemedText.DeprecatedBlack fontWeight={400} fontSize={14} color={theme.textSecondary}>
-                    <Trans>Auto Router API</Trans>
-                  </ThemedText.DeprecatedBlack>
-                  <QuestionHelper text={<Trans>Use the Uniswap Labs API to get faster quotes.</Trans>} />
+                  <Column gap="xs">
+                    <ThemedText.BodySecondary>
+                      <Trans>Auto Router API</Trans>
+                    </ThemedText.BodySecondary>
+                    <ThemedText.Caption color="textSecondary">
+                      <Trans>Finds the best route across liquidity sources.</Trans>
+                    </ThemedText.Caption>
+                  </Column>
                 </RowFixed>
                 <Toggle
                   id="toggle-optimized-router-button"
-                  isActive={!clientSideRouter}
+                  isActive={!clientSideRouterEnabled}
                   toggle={() => {
                     sendEvent({
                       category: 'Routing',
-                      action: clientSideRouter ? 'enable routing API' : 'disable routing API',
+                      action: clientSideRouterEnabled ? 'enable routing API' : 'disable routing API',
                     })
-                    setClientSideRouter(!clientSideRouter)
+                    setClientSideRouterEnabled(!clientSideRouterEnabled)
                   }}
                 />
               </RowBetween>
             )}
-            <RowBetween>
-              <RowFixed>
-                <ThemedText.DeprecatedBlack fontWeight={400} fontSize={14} color={theme.textSecondary}>
-                  <Trans>Expert Mode</Trans>
-                </ThemedText.DeprecatedBlack>
-                <QuestionHelper
-                  text={
-                    <Trans>Allow high price impact trades and skip the confirm screen. Use at your own risk.</Trans>
-                  }
-                />
-              </RowFixed>
-              <Toggle
-                id="toggle-expert-mode-button"
-                isActive={expertMode}
-                toggle={
-                  expertMode
-                    ? () => {
-                        toggleExpertMode()
-                        setShowConfirmation(false)
-                      }
-                    : () => {
-                        toggle()
-                        setShowConfirmation(true)
-                      }
-                }
-              />
-            </RowBetween>
+            <Divider />
+            <ClientSideRouterSettings enabled={clientSideRouterEnabled}>
+              <AutoColumn gap="16px">
+                <MaxSlippageSettings placeholder={placeholderSlippage} />
+                {showDeadlineSettings && (
+                  <>
+                    <Divider />
+                    <TransactionDeadlineSettings />
+                  </>
+                )}
+              </AutoColumn>
+            </ClientSideRouterSettings>
           </AutoColumn>
         </MenuFlyout>
       )}
